requirements.txt
# Task Manager Backend - Python Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
mysql-connector-python==8.2.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Validation
email-validator==2.1.0

# Date/Time
python-dateutil==2.8.2
